JBIDE-25386: Consolidate tests from 'org.hibernate.eclipse.console.test' into new test plugin 'org.jboss.tools.hibernate.orm.test' - Move 'org.hibernate.eclipse.hqleditor.preferences..HQLEditorPreferencePageTest' to 'org.jboss.tools.hibernate.orm.test.HQLEditorPreferencePageTest'
