+++ tests/org.hibernate.eclipse.console.test/META-INF/MANIFEST.MF
@@ -24,7 +24,6 @@ Export-Package: org.hibernate.eclipse.console.test,
  org.hibernate.eclipse.console.test.project,
  org.hibernate.eclipse.console.test.project.xpl,
  org.hibernate.eclipse.console.test.utils,
- org.hibernate.eclipse.hqleditor.preferences,
  org.jmock,
  org.jmock.api,
  org.jmock.integration.junit3,
